+++ src/main/java/com/nhnacademy/store99/bookstore/book_category/controller/BookCategoryController.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.store99.bookstore.book_category.controller;
 
-import com.nhnacademy.store99.bookstore.book.response.BookListElementDTO;
-import com.nhnacademy.store99.bookstore.book_category.response.CategoryParentsDTO;
+import com.nhnacademy.store99.bookstore.book.dto.response.response.BookListElementDTO;
+import com.nhnacademy.store99.bookstore.book_category.dto.response.CategoryParentsDTO;
 import com.nhnacademy.store99.bookstore.book_category.service.BookCategoryService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
