+++ src/main/java/com/nhnacademy/store99/bookstore/book_category/repository/BookCategoryRepository.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.store99.bookstore.book_category.repository;
 
-import com.nhnacademy.store99.bookstore.book.response.BookListElementDTO;
-import com.nhnacademy.store99.bookstore.book_category.response.BookCategoryResponse;
-import com.nhnacademy.store99.bookstore.book_category.response.CategoryParentsDTO;
+import com.nhnacademy.store99.bookstore.book.dto.response.response.BookListElementDTO;
+import com.nhnacademy.store99.bookstore.book_category.dto.response.BookCategoryResponse;
+import com.nhnacademy.store99.bookstore.book_category.dto.response.CategoryParentsDTO;
 import java.util.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
