--- src/main/java/com/nhnacademy/store99/bookstore/book_category/response/BookCategoryResponse.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.store99.bookstore.book_category.response;
-
-public class BookCategoryResponse {
-}
