+++ src/main/java/com/nhnacademy/store99/bookstore/book_image/repository/BookImageRepository.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.store99.bookstore.book_image.repository;
 
-import com.nhnacademy.store99.bookstore.book_image.response.BookImageDTO;
+import com.nhnacademy.store99.bookstore.book_image.dto.response.BookImageDTO;
 
 public interface BookImageRepository {
     BookImageDTO getBookImageData(Long bookId);
