+++ src/main/java/com/nhnacademy/store99/bookstore/book_image/service/BookImageService.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.store99.bookstore.book_image.service;
 
-import com.nhnacademy.store99.bookstore.book_image.response.BookImageDTO;
+import com.nhnacademy.store99.bookstore.book_image.dto.response.BookImageDTO;
 
 public interface BookImageService {
     BookImageDTO getBookImageData(Long bookId);
